icq-backend/pyproject.toml
[tool.poetry]
name = "icq-backend"
version = "0.1.0"
description = ""
authors = ["Johan Mickos <johan.mickos@gmail.com>"]
readme = "README.md"
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^4.8.3"
sklearn = "^0.0.post1"
numpy = "^1.24.2"
scikit-learn = "^1.2.2"
webdriver-manager = "^3.8.5"
pillow = "^9.5.0"
django = "^4.2"
psycopg = "^3.1.8"
psycopg-binary = "^3.1.8"
postgis = "^1.0.4"
gdal = "3.3.2"
scikit-image = "^0.20.0"
fast-colorthief = "^0.0.4"
cairosvg = "^2.7.0"
colorthief = "^0.2.1"
psycopg2-binary = "^2.9.6"

[tool.poetry.dev-dependencies]
pylint = "^2.17.2"
isort = "^5.12.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
black = "^23.3.0"
jupyter = "^1.0.0"
matplotlib = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
